+++ server/services/BpmnToSsotParsing/SsotBaseObjects.js
@@ -1,0 +1,14 @@
+const baseElement = {
+    "type": "",
+    "name": "",
+    "id": "",
+    "predecessorIds": [],
+    "successorIds": []
+}
+
+const metadataObj = {
+    "robotId": "exampleRobot",
+    "starterId": "exampleID"
+}
+
+module.exports = { baseElement, metadataObj  };
