+++ server/services/BpmnToSsotParsing/_tests_/SsotMockForTesting.js
